data_import/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
beautifulsoup4==4.2.1
lxml==3.4.0
prompter==0.3.4
psycopg2==2.5.2
python-dwca-reader==0.6.0
wsgiref==0.1.2
